+++ packages/core/src/slate/editor/getNodeEntries.ts
@@ -1,22 +1,23 @@
-import { Editor } from 'slate';
+import { Editor, EditorNodesOptions } from 'slate';
 import { ENodeMatchOptions, getQueryOptions } from '../../common/queries/match';
 import { Modify } from '../../common/types/utility/types';
-import { ENode } from '../node/TNode';
+import { ENode, TNode } from '../node/TNode';
 import { TNodeEntry } from '../node/TNodeEntry';
 import { TEditor, Value } from './TEditor';
 import { unhangRange, UnhangRangeOptions } from './unhangRange';
 
 export type GetNodeEntriesOptions<V extends Value = Value> = Modify<
-  NonNullable<Parameters<typeof Editor.nodes>[1]>,
+  NonNullable<EditorNodesOptions<TNode>>,
   ENodeMatchOptions<V>
->;
+> &
+  UnhangRangeOptions;
 
 /**
  * Iterate through all of the nodes in the Editor.
  */
 export const getNodeEntries = <N extends ENode<V>, V extends Value = Value>(
   editor: TEditor<V>,
-  options?: GetNodeEntriesOptions<V> & UnhangRangeOptions
+  options?: GetNodeEntriesOptions<V>
 ): Generator<TNodeEntry<N>, void, undefined> => {
   unhangRange(editor, options?.at, options);