+++ packages/nodes/mention/src/handlers/mentionOnKeyDownHandler.spec.tsx
@@ -1,6 +1,7 @@
 /** @jsx jsx */
 
 import { jsx } from '@udecode/plate-test-utils';
+import * as isHotkey from 'is-hotkey';
 import { createEditorWithMentions } from '../__tests__/createEditorWithMentions';
 import { mentionOnKeyDownHandler } from './mentionOnKeyDownHandler';
 
@@ -22,6 +23,8 @@ describe('mentionOnKeyDownHandler', () => {
     );
 
     const plugin = editor.pluginsByKey.mention;
+
+    jest.spyOn(isHotkey, 'default').mockReturnValue(true);
 
     mentionOnKeyDownHandler()(editor)(
       new KeyboardEvent('keydown', { key: 'Escape' }) as any