+++ packages/nodes/mention/src/handlers/moveSelectionByOffset.ts
@@ -1,4 +1,5 @@
 import { moveSelection, PlateEditor, Value } from '@udecode/plate-core';
+import isHotkey from 'is-hotkey';
 import { Range } from 'slate';
 import { KeyboardEventHandler } from './KeyboardEventHandler';
 
@@ -19,13 +20,13 @@ export const moveSelectionByOffset: <V extends Value>(
     return false;
   }
 
-  if (event.key === 'ArrowLeft') {
+  if (isHotkey('left', event)) {
     event.preventDefault();
     moveSelection(editor, { unit: 'offset', reverse: true });
     return true;
   }
 
-  if (event.key === 'ArrowRight') {
+  if (isHotkey('right', event)) {
     event.preventDefault();
     moveSelection(editor, { unit: 'offset' });
     return true;