+++ packages/nodes/horizontal-rule/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-horizontal-rule
 
+## 11.2.0
+
 ## 11.1.0
 
 ## 11.0.6